Add IEntityCollection<TEntity> for working with entity collections
